Load config options from a file to an importable dict.
If the config file doesn't exist, create a default one
from a dict.

Likewise, create a default users file from a dict if
it is missing (instead of copying a whole file).

Make users_file an attribute of Timesheet objects.

Move most code from __main__.py to chronophore.py so it's
easier to test.